flask/requirements.txt
pip==22.2.2
wheel==0.36.2
Brotli==1.0.9
click==8.0.3
dash==2.6.1
dash-bootstrap-components
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-renderer
dash-table==5.0.0
Flask==2.2.2
Flask-Compress==1.12.0
future==0.18.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.0
pandas==1.4.3
plotly==5.0.0
altair==4.2.0
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
setuptools==57.5.0
six==1.16.0
Werkzeug==2.2.2
gunicorn==20.0.4
Flask-SQLAlchemy==2.5.1
psycopg2-binary==2.8.6